Enable swap and undo files again
The new settings protect against:

1. A crash during editing (put swap file in home directory so it won't
   be tracked in version control systems)
2. Editing the same file with 2 instances of Vim (where one Vim can
   overwrite the changes of another Vim instance)
3. A crash during the save process itself
4. Saving new file contents but wanting the old version back

Version control can only protect against case #4.

These directories need to be pre-made:
~/swap/
~/backup/
~/undo/

Also, the files are left in the temporary directories, so disk space may
be an issue if the swap files pile up and fill up the disk space
(probably won't be an issue when editing text files, which tend to be
quite small).

https://begriffs.com/posts/2019-07-19-history-use-vim.html#backups-and-undo