Swap Markdown plugins, use Previewer
Unfortunately, plasticboy's vim-markdown is incompatible with vim-one
colorscheme:

 * https://github.com/plasticboy/vim-markdown/issues/461

The symptom was that headers marked with "#" and "##" were not being
highlighted with color.

While plasticboy's vim-markdown is more featureful because it allows
section folding, I decided to switch back to garbrielelana's
vim-markdown because it handles syntax highlighting better.
Unfortunately, it doesn't have section folding.

---

I also installed a new plugin, iamcco's markdown-preview, which does an
awesome browser live preview with live scrolling of the markdown file
inside NeoVim. It renders the markdown like GitHub, and seems to be even
better than Grip, which I was using before:

 * https://github.com/joeyespo/grip
 * https://github.com/iamcco/markdown-preview.nvim