Improve triggers for search word/word under cursor
For some reason, having "<C-u>" for "ctrl-u to delete everything before
cusor" is required for to trigger the process for "search for a word
that you can begin typing", but the same ctrl-u interferes with the
functionality for "begin searching for word under cursor".

Somehow, "Ctrl-U" moves the cursor to the beginning of the file for some
reason, and the resulting search starts at the first line of the file
and not at the word under the cursor.