Use leader-H, add config for markdown-preview
Leader-H will show git hunk diffs on lines that have a git-signify
symbol in the gutter.

Still trying to get markdown-preview to work by allowing other clients
on the local network to connect to the server (not sure if it's working
yet or if the plugin got broken during an automatic update of CoC).