Turn off scrolloff, interferes with scrolling
ctrl-y to scroll window down one line and
ctrl-e to scroll window up one line

while both of them preserving the cursor on the line position; these
2 functions were disrupted by the scrolloff=2 setting, which
I originally set so if I was on the last line of the file, it didn't
seem like I was on the very edge of the terminal emulator (like iTerm or Blink iOS Shell).

However, it was more often the case that I wanted to move the current
line to the top of the current window view, and the scrolloff would not
actually produce the line at the top but make it the 3rd line at the
top.