Fix bad visual bug with ctrl-c & floating windows
Finally found the answer to using ctrl-c to close the CoC floating menu.
Sometimes, a line from the floating menu will permanently remain and cause visual
clutter on the screen.

Chemzqm says the reliance of `getchar` from Vim causes this behavior, so
we need to explicity call the exit procedure for the ctrl-c binding.

https://github.com/neoclide/coc.nvim/issues/639