Defer startup of NERDTree again
I don't like the startup delay, so I'm disabling NERDTree on startup
again.

I'll accept the weird behavior of "nvim ." and netrw starting up with a strange buffer list in lightline-bufferline (shows a temporary buffer in the tab bar for netrw) in order to speed up opening files with Neovim.