Use spaces again instead of tabs
The main purpose of using tabs is for increased accessibility: you can
cusotmize the width of the tab.

However, GitHub doesn't recognize my desire to set 2 spaces for tabs
inside '.editorconfig', so GitHub always renders code blocks with
a gigantic span of 8 spaces for tabs, which makes code difficult to read
and removes precious horizontal space for code.