Fix command search, file save, search highlighting
Command Search:
Enter was remapped to 'clear search highlighting'. However, this caused the command search history window 'q:' to block Enter as 'execute
current line'. Fix was to remap the unused '\' key to 'clear search highlighting'.

File Save:
Having comments at the end of the current line with the " symbol wasn't causing problems in the rest of the init.nvim file. However, for the
specific case of using leader-w to initiate file save, this caused the cursor to jump forward the same number of times as there was spaces
before the " symbol." Fix was to move the comment to the above line.

Search Highlighting:
Pressing 'Enter' during q: command search window was disabled because
Enter was remapped to 'clear search highlighting'. Fix was to remap the
unused '\' key to clear search highlights. (The default leader key is
'\', but I already remapped that to 'Space'.)