Fix Tmux copy and paste
Create "leader-l" to toggle line numbers and the gutter (signcolumn for
git change symbols).

When selecting text in Vim or selecting with the mouse, these operations only work inside Vim. You're not able to 'yank' to the system clipboard.

To help, it's better to turn off line numbers and the gutter so you can 'option-select' with the mouse or use Tmux copy mode to copy multiple lines without having the line numbers screw up the beginning of the code lines.

Tmux copy mode 'ctrl-x [' will copy to the system clipboard (using the
utility 'reattach-to-user-namespace').