Map leader-q to 'save buffer w/o kill window'
Tried to use this native Vim method:
:lclose<bar>b#<bar>bd #<CR>

From Stackoverflow:
https://stackoverflow.com/a/29179159

But it seems to keep the old buffer in the jump list, so it was possible
to accidentally go back to it with ctrl-o to re-open it, or it caused
strange temporary buffers to remain open.

I'm using the plugin 'vim-bbye' instead.